requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
parrot==1.0.0
spacy==3.7.2
torch>=2.0.0
transformers>=4.30.0
protobuf>=3.20.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

